+++ src/aks-preview/azext_aks_preview/_help.py
@@ -59,6 +59,12 @@ helps['aks create'] = """
         - name: --admin-username -u
           type: string
           short-summary: User account to create on node VMs for SSH access.
+        - name: --windows-admin-username
+          type: string
+          short-summary: User account to create on windows node VMs.
+        - name: --windows-admin-password
+          type: string
+          short-summary: User account password to use on windows node VMs.
         - name: --aad-client-app-id
           type: string
           short-summary: (PREVIEW) The ID of an Azure Active Directory client application of type "Native". This
@@ -83,6 +89,10 @@ helps['aks create'] = """
           short-summary: A specific IP address and netmask for the Docker bridge, using standard CIDR notation.
           long-summary: This address must not be in any Subnet IP ranges, or the Kubernetes service address range.
                         For example, 172.17.0.1/16.
+        - name: --load-balancer-sku
+          type: string
+          short-summary: Azure Load Balancer SKU selection for your cluster. Basic or Standard.
+          long-summary: Select between Basic or Standard Azure Load Balancer SKU for your AKS cluster.
         - name: --enable-addons -a
           type: string
           short-summary: Enable the Kubernetes addons in a comma-separated list.
@@ -90,6 +100,8 @@ helps['aks create'] = """
             These addons are available:
                 http_application_routing - configure ingress with automatic public DNS name creation.
                 monitoring - turn on Log Analytics monitoring. Uses the Log Analytics Default Workspace if it exists, else creates one. Specify "--workspace-resource-id" to use an existing workspace.
+                virtual-node - enable AKS Virtual Node (PREVIEW). Requires --subnet-name to provide the name of an existing subnet for the Virtual Node to use.
+                azure-policy - enable Azure policy (PREVIEW).
         - name: --disable-rbac
           type: bool
           short-summary: Disable Kubernetes Role-Based Access Control.
@@ -142,6 +154,12 @@ helps['aks create'] = """
         - name: --enable-vmss
           type: bool
           short-summary: (PREVIEW) Enable VMSS agent type.
+        - name: --enable-pod-security-policy
+          type: bool
+          short-summary: (PREVIEW) Enable pod security policy.
+        - name: --node-resource-group
+          type: string
+          short-summary: The node resource group is the resource group where all customer's resources will be created in, such as virtual machines.
     examples:
         - name: Create a Kubernetes cluster with an existing SSH public key.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --ssh-key-value /path/to/publickey
@@ -197,6 +215,12 @@ helps['aks update'] = """
         - name: --api-server-authorized-ip-ranges
           type: str
           short-summary: List of authorized IP ranges (separated by comma) for apiserver. Set to "" for disabling it.
+        - name: --enable-pod-security-policy
+          type: bool
+          short-summary: (PREVIEW) Enable pod security policy.
+        - name: --disable-pod-security-policy
+          type: bool
+          short-summary: (PREVIEW) Disable pod security policy.
     examples:
       - name: Enable cluster-autoscaler within node count range [1,5]
         text: az aks update --enable-cluster-autoscaler --min-count 1 --max-count 5 -g MyResourceGroup -n MyManagedCluster
@@ -206,6 +230,10 @@ helps['aks update'] = """
         text: az aks update --update-cluster-autoscaler --min-count 1 --max-count 10 -g MyResourceGroup -n MyManagedCluster
       - name: Enable authorized IP ranges for apiserver.
         text: az aks update --api-server-authorized-ip-ranges 172.0.0.10/16,168.10.0.10/18 -g MyResourceGroup -n MyManagedCluster
+      - name: Enable pod security policy.
+        text: az aks update --enable-pod-security-policy -g MyResourceGroup -n MyManagedCluster
+      - name: Disable pod security policy.
+        text: az aks update --disable-pod-security-policy -g MyResourceGroup -n MyManagedCluster
 """
 
 helps['aks nodepool'] = """
@@ -254,6 +282,15 @@ helps['aks nodepool add'] = """
         - name: --os-type
           type: string
           short-summary: The OS Type. Linux or Windows.
+        - name: --enable-cluster-autoscaler -e
+          type: bool
+          short-summary: Enable cluster autoscaler.
+        - name: --min-count
+          type: int
+          short-summary: Minimun nodes count used for auto scaler, when "--enable-cluster-autoscaler" specified. Please specifying the value in the range of [1, 100]
+        - name: --max-count
+          type: int
+          short-summary: Maximum nodes count used for auto scaler, when "--enable-cluster-autoscaler" specified. Please specifying the value in the range of [1, 100]
 """
 
 helps['aks nodepool scale'] = """
@@ -274,7 +311,69 @@ helps['aks nodepool upgrade'] = """
           short-summary: Version of Kubernetes to upgrade the node pool to, such as "1.11.12".
 """
 
+helps['aks nodepool update'] = """
+    type: command
+    short-summary: Update a node pool to enable/disable cluster-autoscaler or change min-count or max-count
+    parameters:
+        - name: --enable-cluster-autoscaler -e
+          type: bool
+          short-summary: Enable cluster autoscaler.
+        - name: --disable-cluster-autoscaler -d
+          type: bool
+          short-summary: Disable cluster autoscaler.
+        - name: --update-cluster-autoscaler -u
+          type: bool
+          short-summary: Update min-count or max-count for cluser auto-scaler.
+        - name: --min-count
+          type: int
+          short-summary: Minimun nodes count used for auto scaler, when "--enable-cluster-autoscaler" specified. Please specifying the value in the range of [1, 100]
+        - name: --max-count
+          type: int
+          short-summary: Maximum nodes count used for auto scaler, when "--enable-cluster-autoscaler" specified. Please specifying the value in the range of [1, 100]
+    examples:
+      - name: Enable cluster-autoscaler within node count range [1,5]
+        text: az aks nodepool update --enable-cluster-autoscaler --min-count 1 --max-count 5 -g MyResourceGroup -n nodepool1 --cluster-name MyManagedCluster
+      - name: Disable cluster-autoscaler for an existing cluster
+        text: az aks nodepool update --disable-cluster-autoscaler -g MyResourceGroup -n nodepool1 --cluster-name MyManagedCluster
+      - name: Update min-count or max-count for cluster auto-scaler.
+        text: az aks nodepool update --update-cluster-autoscaler --min-count 1 --max-count 10 -g MyResourceGroup -n nodepool1 --cluster-name MyManagedCluster
+"""
+
 helps['aks nodepool delete'] = """
     type: command
     short-summary: Delete the agent pool in the managed Kubernetes cluster.
+"""
+
+helps['aks enable-addons'] = """
+type: command
+short-summary: Enable Kubernetes addons.
+long-summary: |-
+    These addons are available:
+        http_application_routing - configure ingress with automatic public DNS name creation.
+        monitoring - turn on Log Analytics monitoring. Requires "--workspace-resource-id".
+        virtual-node - enable AKS Virtual Node (PREVIEW). Requires "--subnet-name".
+        azure-policy - enable Azure policy (PREVIEW).
+parameters:
+  - name: --addons -a
+    type: string
+    short-summary: Enable the Kubernetes addons in a comma-separated list.
+  - name: --workspace-resource-id
+    type: string
+    short-summary: The resource ID of an existing Log Analytics Workspace to use for storing monitoring data.
+  - name: --subnet-name -s
+    type: string
+    short-summary: The subnet name for the virtual node to use.
+examples:
+  - name: Enable Kubernetes addons. (autogenerated)
+    text: az aks enable-addons --addons virtual-node --name MyManagedCluster --resource-group MyResourceGroup --subnet-name VirtualNodeSubnet
+    crafted: true
+"""
+
+helps['aks get-versions'] = """
+type: command
+short-summary: Get the versions available for creating a managed Kubernetes cluster.
+examples:
+  - name: Get the versions available for creating a managed Kubernetes cluster
+    text: az aks get-versions --location westus2
+    crafted: true
 """