+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,30 @@
 
 Release History
 ===============
+0.4.4
++++++
+* Add support for per node pool auto scaler settings.
+* Add `az aks nodepool update` to allow users to change auto scaler settings per node pool.
+* Add support for Standard sku load balancer.
+
+0.4.1
++++++
+* Add `az aks get-versions -l location` to allow users to see all managed cluster versions.
+* Add `az aks get-upgrades` to get all available versions to upgrade.
+* Add '(preview)' suffix if kubernetes version is preview when using `get-versions` and `get-upgrades`
+
+0.4.0
++++++
+* Add support for Azure policy add-on.
+
+0.3.2
++++++
+* Add support of customizing node resource group
+
+0.3.1
++++++
+* Add support of pod security policy.
+
 0.3.0
 +++++
 * Add support of feature `--node-zones`